chore: refactor dots pagination container, fix last active style dot
